// src-angular/src/app/modules/core/models/developer.model.ts
export interface DeveloperResponse {
  id: number;
  name: string;
  addressCountry: string;
  addressStreet: string;
  addressBuildingNumber: string;
  addressFlatNumber: string;
  addressPostalCode: string;
  telephoneNumber: string;
  faxNumber: string;
  email: string;
  taxIdentificationNumber: string;
  cityId: number;
  logoUrl: string;
  code: string;
  createdAt: string;
  updatedAt: string;
  deletedAt: string;
}